+++ mobile/android/app/src/main/kotlin/com/walkingrpg/walking_rpg_mobile/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.walkingrpg.walking_rpg_mobile
+
+import io.flutter.embedding.android.FlutterFragmentActivity
+
+class MainActivity : FlutterFragmentActivity()